Move Method: prerequisite-PR-creation cluster -> repair_body.py
prerequisite_branch_name, prerequisite_title, prerequisite_body, and
create_repair_prerequisite move out of AdminBypassRepairer into
repair_body.py. repair_check's one call site is re-pointed in this same
commit, passing self.gh/self.ledger/self.logger/self.repo explicitly now
that they're no longer available implicitly through self.

Clustered as one slice: the three naming helpers have no caller except
create_repair_prerequisite. Depends on the two previous slices
(create_repair_prerequisite calls git_output/hard_reset_work_root and
validate_current_pr_body). push_branch's one call site inside
create_repair_prerequisite is inlined to call safe_push directly (the
same underlying call push_branch already made, since push_branch itself
stays a thin method on AdminBypassRepairer and isn't moving).

No behavior change, verified by the full existing test suite passing with
only mechanical mock-target updates.
